% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{add_value_labels}
\alias{add_value_labels}
\title{Add Value Labels}
\usage{
add_value_labels(x, y, labels = y, cex = 0.9)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{labels}{labels, default: y coordinates}

\item{cex}{character expansion factor, default: 0.9}
}
\description{
Add Value Labels
}
