% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{create_and_plot_rain_events}
\alias{create_and_plot_rain_events}
\title{plot rain events from given rain data}
\usage{
create_and_plot_rain_events(
  frmRain,
  strTimestamp = "Zeitstempel",
  strPdf = NULL,
  dbg = FALSE
)
}
\arguments{
\item{frmRain}{data frame containing rain data}

\item{strTimestamp}{name of timestamp field, default: "Zeitstempel"}

\item{strPdf}{optional. full path to pdf output file}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
plot rain events from given rain data.
  Rain data is filtered for rows where signals from all gauges are available 
  and where the sum of signals is greater than zero.
}
