% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDataSource.R
\name{defaultDictionary}
\alias{defaultDictionary}
\title{Get Default Dictionary from Meta Database}
\usage{
defaultDictionary(mdb = mmdb(), dbg = FALSE)
}
\arguments{
\item{mdb}{full path to meta database, default: \code{\link{mmdb}()}}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
Get a default dictionary from the meta database
}
