% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibGeneral.R
\name{documentPackageFunctionDependencies}
\alias{documentPackageFunctionDependencies}
\title{plot graphs showing the dependencies between package functions}
\usage{
documentPackageFunctionDependencies(packagenames, to.pdf = FALSE)
}
\arguments{
\item{packagenames}{vector of character containing the names of the packages of which
functional dependencies are to be documented.}

\item{to.pdf}{if TRUE, graphical output is written to a pdf file}
}
\description{
plot graphs showing the dependencies between package functions
}
\examples{
# Show names of all installed KWB-packages
grep("^kwb\\\\.", library()$results[, "Package"], value = TRUE)
  
# Document one of the installed packages
\dontrun{
documentPackageFunctionDependencies("kwb.plot")
}
}
