% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{example_rain_data}
\alias{example_rain_data}
\title{example rain data}
\usage{
example_rain_data(version = 1)
}
\arguments{
\item{version}{Number representing version of data to return. Default: 1}
}
\description{
example rain data
}
