% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{get_rain_stat}
\alias{get_rain_stat}
\title{Get Rain Stats}
\usage{
get_rain_stat(rain, strTimestamp = names(rain)[1])
}
\arguments{
\item{rain}{data frame containing rain data}

\item{strTimestamp}{Name of column containing timestamps. Default: Name of
first column in \code{rain}}
}
\description{
Get Rain Stats
}
