% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsAllCoefAnaCombis}
\alias{hsAllCoefAnaCombis}
\title{hsAllCoefAnaCombis}
\usage{
hsAllCoefAnaCombis(n, dbg.level = n)
}
\arguments{
\item{n}{number of elements to be combined}

\item{dbg.level}{if not 0, combinations are shown when dbg.level-th element just changed}
}
\description{
Generates all possible combinations of events as used for coefficient
  analysis
}
