% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsAllCombis}
\alias{hsAllCombis}
\title{hsAllCombis}
\usage{
hsAllCombis(x)
}
\arguments{
\item{x}{vector of elements of which to create combinations}
}
\description{
Generate all possible combinations of elements in \code{x} with order mattering
}
