% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibTimeshift.R
\name{hsAllTimeshiftPlots}
\alias{hsAllTimeshiftPlots}
\title{All Timeshift Plots}
\usage{
hsAllTimeshiftPlots(
  frmOrig,
  frmUs,
  frmDs,
  evt,
  fieldNames,
  plusOverview = FALSE,
  type1 = "l",
  fracts = c(0, 0.75, 0.5, 0.25, 0.1, 0.05),
  fieldPrefix = ""
)
}
\arguments{
\item{frmOrig}{Original time-series}

\item{frmUs}{upstream shifted time-series}

\item{frmDs}{downstream shifted time-series}

\item{evt}{event list}

\item{fieldNames}{vector containing the relevant table field names as named elements:
tsOrig: timestamp in Q time-series), 
tsUsDs: timestamp in us/ds time-series),
fieldQ: name of field containing Q,
fieldP: name of field containing wq parameter}

\item{plusOverview}{If true, one plot comprising the whole time-series in \code{frmOrig} is added}

\item{type1}{Plot type (default: "l" = line)}

\item{fracts}{Vector of scaling factors. For each factor a plot is generated representing
the corresponding fraction of the whole time interval of the event. 
A factor of zero will plot the whole event.}

\item{fieldPrefix}{Default: ""}
}
\description{
For one water quality parameter, all overflow events given in "evt" are
plotted in different scales given by "fracts" (fractions of interval length).
The time-series of the water quality parameter at the container, upstream and
downstream are plotted over time as well as the flow.
}
