% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibMiaCsoData.R
\name{hsAttribMatrix}
\alias{hsAttribMatrix}
\title{Attribute Strings to Matrix}
\usage{
hsAttribMatrix(attribs)
}
\arguments{
\item{attribs}{Vector containing strings of the form "<key1>=<val1>,<key2>=<val2>,..."}
}
\description{
Converts a vector of attribute strings to a matrix with as many columns as
  there are different attributes occurring in the vector and each row
  representing an element of the vector for which the values of assigned
  attributes will occur in the corresponding attribute column.
}
