% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsBrowseCoefAnaRes}
\alias{hsBrowseCoefAnaRes}
\title{hsBrowseCoefAnaRes}
\usage{
hsBrowseCoefAnaRes(tree, combilen = -1)
}
\arguments{
\item{tree}{list representing a \code{tree} structure as returned by \code{\link{hsCoefAna}}}

\item{combilen}{length of combinations. Default: -1}
}
\description{
Browse through result \code{tree} of regression coefficient analysis and "rbind"
  data frames \emph{linreg}
}
