% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsBrowseCoefAnaResList}
\alias{hsBrowseCoefAnaResList}
\title{hsBrowseCoefAnaResList}
\usage{
hsBrowseCoefAnaResList(reslist, dbg.level = 10)
}
\arguments{
\item{reslist}{result list as returned by hsCoefAna(..., recursive = FALSE, aslist = TRUE))}

\item{dbg.level}{debug level}
}
\description{
Browse through result tree of regression coefficient analysis and "rbind"
  data frames \emph{linreg}
}
