% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsBrowseCombis}
\alias{hsBrowseCombis}
\title{Browse Combinations}
\usage{
hsBrowseCombis(tree, combis = list())
}
\arguments{
\item{tree}{list representing a \code{tree} structure as returned by \code{\link{hsCoefAna}}}

\item{combis}{List of combinations. Default: \code{list()}}
}
\value{
list with first element containing matrix of combinations of length 1, 
  second element containing matrix of combinations of length 2, and so on.
}
\description{
browses through result \code{tree} and collects all combinations
}
