% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCalib.R
\name{hsCalibAna}
\alias{hsCalibAna}
\title{Analyse Calibration Uncertainty}
\usage{
hsCalibAna(data, dbg = TRUE, doplot = TRUE, plot.main = "", pdf)
}
\arguments{
\item{data}{data frame with column \code{evtid} (event ID) and further
columns required by \code{\link{hsTestCombi}} to which the data frame
is passed.}

\item{dbg}{If \code{TRUE}, debug messages are shown}

\item{doplot}{logical telling whether to plot within
\code{\link{hsTestCombi}}. Default: \code{TRUE}}

\item{plot.main}{plot title. Default: ""}

\item{pdf}{path to PDF file}
}
\description{
Try all different combinations of events for calibration
}
