% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsCoefAna}
\alias{hsCoefAna}
\title{hsCoefAna}
\usage{
hsCoefAna(
  data,
  recursive = TRUE,
  evtNums = unique(data$evtID),
  aslist = recursive,
  uselm = FALSE,
  prep = FALSE,
  ...,
  dbg.level = max(2, length(evtNums) - 8)
)
}
\arguments{
\item{data}{\code{data} frame containing columns \emph{tstamp} (time stamp), \emph{pval}
(probe value), \emph{lval} (lab value), \emph{evtID} (event ID)}

\item{recursive}{if TRUE, the \code{recursive} version hsCoefAnaRes of the regression coefficient 
analysis is used, otherwise the non-recursive version. default: TRUE}

\item{evtNums}{event numbers to be considered for the analysis. Only considered when
\emph{recursive} == TRUE. default: all distinct values provided in 
column \emph{evtID} of \emph{data})}

\item{aslist}{default: boolean value given in \emph{recursive}}

\item{uselm}{if TRUE, the lm function is used to calculate the linear regression,
otherwise (\code{uselm} == FALSE) the regression is calculated "manually" 
which is much faster. Default: FALSE}

\item{prep}{Default: FALSE}

\item{\dots}{further arguments passed to \code{\link{hsCombiLinReg}}, e.g. \emph{clever}}

\item{dbg.level}{default: max(2, length(\code{evtNums}) - 8)}
}
\description{
regression coefficient analysis
}
