% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsCoefAnaNonRec}
\alias{hsCoefAnaNonRec}
\title{hsCoefAnaNonRec}
\usage{
hsCoefAnaNonRec(data, uselm = FALSE, aslist = FALSE, ..., dbg.level = 1)
}
\arguments{
\item{data}{\code{data} frame containing columns \emph{tstamp} (time stamp), \emph{pval}
(probe value), \emph{lval} (lab value), \emph{evtID} (event ID)}

\item{uselm}{if TRUE, the lm function is used to calculate the linear regression,
otherwise (\code{uselm} == FALSE) the regression is calculated "manually" 
which is much faster. default: FALSE}

\item{aslist}{if TRUE the result is retunred in forms of a list with each list element
representing one combination. otherwise in forms of a database with
columns \emph{np} (number of points), \emph{offset}, \emph{slope},
\emph{combi}. Default: FALSE.}

\item{\dots}{further arguments passed to \code{\link{hsCombiLinReg}}, e.g. \emph{clever}}

\item{dbg.level}{debug level}
}
\description{
non-recursive version of regression coefficient analysis
}
