% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsCoefAnaRec}
\alias{hsCoefAnaRec}
\title{hsCoefAnaRec}
\usage{
hsCoefAnaRec(
  data,
  evtNums = unique(data$evtID),
  combi = NULL,
  tree = TRUE,
  resframe = NULL,
  uselm = FALSE,
  ...,
  dbg.level = 1
)
}
\arguments{
\item{data}{\code{data} frame containing columns \emph{tstamp} (time stamp), \emph{pval}
(probe value), \emph{lval} (lab value), \emph{evtID} (event ID)}

\item{evtNums}{event numbers to be considered for the analysis (default: all distinct
values provided in column \emph{evtID} of \emph{data})}

\item{combi}{current combination to be evaluated and to be the base for the next 
combinations to be determined}

\item{tree}{if TRUE, result is given in a \code{tree} structure, otherwise as a \code{data} frame}

\item{resframe}{if \emph{tree} is FALSE, this argument contains the results that have
been found so far in a \code{data} frame}

\item{uselm}{if TRUE, the lm function is used to calculate the linear regression,
otherwise (\code{uselm} == FALSE) the regression is calculated "manually" 
which is much faster. default: FALSE}

\item{\dots}{further arguments passed to \code{\link{hsCombiLinReg}}, e.g. \emph{clever}}

\item{dbg.level}{debug level}
}
\value{
Recursive list representing a \code{tree} structure. At the top level the list
  contains elements \emph{e<i>} where <i> are the event IDs to be considered 
  (elements in \emph{evtNums}).
  The sub lists below the top level (but not the "leafs" of the \code{tree}) also 
  contain elements \emph{e<j>} where <j> are the "remaining" event IDs, 
  i.e. the IDs that do not yet occur in the "path" of event IDs
  leading to the respective sub \code{tree}. These sub lists also have elements
  \emph{combi} (vector of event IDs representing the respective event 
  combination) and \emph{linreg} containing the results from linear 
  regression. In fact, \emph{linreg} is a \code{data} frame with each line 
  representing the \emph{slope} and \emph{offset} of the linear regression 
  through \emph{np} number of points, taken from the events in \emph{combi}.
}
\description{
recursive version of regression coefficient analysis
}
