% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsCombiLinReg}
\alias{hsCombiLinReg}
\title{Linear Regression for Event Combination}
\usage{
hsCombiLinReg(
  data,
  combi,
  uselm = FALSE,
  clever = FALSE,
  prep = FALSE,
  calc.rmse = TRUE,
  dbg = FALSE
)
}
\arguments{
\item{data}{\code{data} frame containing columns \emph{tstamp} (time stamp), \emph{pval}
(probe value), \emph{lval} (lab value), \emph{evtID} (event ID)}

\item{combi}{combination of events for which linear regressions are to be calculated
in the following way: the first event numbers in \code{combi}, at positions 
\code{seq_len(length(combi) - 1)}, are considered to be "base" events, i.e. events
of which all \code{data} points are considered for the linear regression.  
The \code{data} points belonging to the event given at the last position
of \code{combi} are added "point by point" to these "base points" and each time
a separate regression is calculated}

\item{uselm}{if TRUE, the lm function is used to calculate the linear regression,
otherwise (\code{uselm} == FALSE) the regression is calculated "manually" 
which is much faster. default: FALSE}

\item{clever}{if TRUE, sums and means are updated by knowledge of previous values with
the current \code{data} point, otherwise they are always recalculated for all
datapoints to be considered}

\item{prep}{if TRUE, \code{data} is expected to contain columns \emph{x2} (squares of x),
\emph{xy} (product of x and y values). Unfortunately, this does not give
a better performance...}

\item{calc.rmse}{if TRUE, the root mean square error (RMSE) is calculated}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
Calculation of linear regressions for given combination of events
}
