% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCalib.R
\name{hsCurCal}
\alias{hsCurCal}
\title{Current calibration}
\usage{
hsCurCal(moniPoint, parAcronym, globOnly = FALSE, mdbCal = NULL)
}
\arguments{
\item{moniPoint}{acronym of monitoring point}

\item{parAcronym}{acronym of parameter}

\item{globOnly}{logical. Globals only? Default: FALSE}

\item{mdbCal}{Path to database containing queries that getting calibrated
data according to the currently active calibration setting}
}
\description{
Get calibrated data according to current calibration
}
