% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDataAvailability.R
\name{hsDataAvailability.old}
\alias{hsDataAvailability.old}
\title{Data availability of time series data}
\usage{
hsDataAvailability.old(
  info,
  dateFirst = NULL,
  dateLast = NULL,
  tstep = NULL,
  dbg = FALSE
)
}
\arguments{
\item{info}{list with the named elements (\emph{mdb}: full path to Access database, 
\emph{tbl}: table name, \emph{tsField}: name of timestamp field, 
\emph{parField}: name of parameter field)}

\item{dateFirst}{Date object representing first date to be considered}

\item{dateLast}{Date object representing last date to be considered}

\item{tstep}{expected time step between time stamps in seconds. Default: minimum 
time difference found between consecutive timestamps in given interval}

\item{dbg}{If TRUE, debug messages will be shown}
}
\value{
data.frame with each row representing a day within the specified time
  interval and columns \eqn{intervalBeg} (day), \eqn{n<Par>} 
  (number of non-NA-values in column <Par> within the interval)
  and \eqn{p<Par>} (data availability of parameter <Par> in percent 
  = number of available non-NA-values divided by maximum possible 
  number of non-NA-values per day (= 86400 / \emph{tstep}).
}
\description{
Data availability of time series data
}
