% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDataAvailability.R
\name{hsDataAvailability}
\alias{hsDataAvailability}
\title{Data Availability in time-series data}
\usage{
hsDataAvailability(
  data,
  tstep = minTimeStep(data[[1]], dbg = dbg),
  interval = 60 * 60 * 24,
  includeCount = TRUE,
  dbg = FALSE
)
}
\arguments{
\item{data}{\code{data} frame with timestamp in first column}

\item{tstep}{expected timestep between consecutive timestamps in seconds. Default:
minimum time difference occurring in timestamps of \emph{data}.}

\item{interval}{length of time intervals to which \code{data} is grouped, in seconds. Default:
60*60*24 = one day intervals; \code{data} availability is calculated separately
for each time \code{interval}.}

\item{includeCount}{if TRUE, not only the \code{data} availability in percent but also the number
of records per \code{interval} from which the percentage has been calculated
are included as separate columns in the result \code{data} frame.}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
\code{data} availability in time-series \code{data}
}
