% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibMiaCsoData.R
\name{hsDirStructure}
\alias{hsDirStructure}
\title{Read Directory Structure from Metadata Database}
\usage{
hsDirStructure(asMatrix = FALSE, dbg = FALSE)
}
\arguments{
\item{asMatrix}{logical indicating whether to return a matrix or not}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
Gets recursively defined directory structure from RMeta.mdb
}
