% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibGeneral.R
\name{hsDropExistingTable}
\alias{hsDropExistingTable}
\title{Drop an existing table (user interaction)}
\usage{
hsDropExistingTable(channel, strTable, boolAsk = TRUE)
}
\arguments{
\item{channel}{database connection}

\item{strTable}{table name}

\item{boolAsk}{logical indication whether the user should be asked before 
dropping the table}
}
\value{
table name of created table.
}
\description{
Drop an existing table (user interaction)
}
