% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsEqualUntilPos}
\alias{hsEqualUntilPos}
\title{hsEqualUntilPos}
\usage{
hsEqualUntilPos(combi1, combi2)
}
\arguments{
\item{combi1}{vector}

\item{combi2}{vector of same mode as \code{combi1}}
}
\description{
returns the first index at which elements in \code{combi1} and \code{combi2} differ
}
