% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsExampleCoefData}
\alias{hsExampleCoefData}
\title{hsExampleCoefData}
\usage{
hsExampleCoefData(
  nevts = 5,
  step = 30,
  ex.avail = FALSE,
  dev = 0.01,
  dbg = FALSE
)
}
\arguments{
\item{nevts}{number of gaps to be produced in continuous examle data in order to split
events. default: 5}

\item{step}{timestep in seconds. default: 30}

\item{ex.avail}{example data available? default: FALSE}

\item{dev}{deviation. default: 0.01}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
provide example dataset for coefficient analysis with \code{\link{hsCoefAna}}
}
