% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDataSource.R
\name{hsFileCands}
\alias{hsFileCands}
\title{File Candidates}
\usage{
hsFileCands(mdb, dict.lst, dbg = FALSE)
}
\arguments{
\item{mdb}{path to MS Access database}

\item{dict.lst}{path dictionary (a list)}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
find file candidates according to paths defined in dictionary
}
