% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibImpSenData.R
\name{hsGetOrCreateParID}
\alias{hsGetOrCreateParID}
\title{hsGetOrCreateParID}
\usage{
hsGetOrCreateParID(parInfo, parDefs)
}
\arguments{
\item{parInfo}{parameter definition}

\item{parDefs}{list of parameter definitions}
}
\value{
List with two elements \emph{myParID} and \emph{parDefs}.
  If the parameter defined in \emph{parInfo} was found in the list of all 
  parameters \emph{parDefs} that have been found so far, the unique
  parameter ID used in \code{parDefs} is returned in \emph{myParID}. If the
  parameter was not found it is added to \code{parDefs} and
}
\description{
Lookup a water quality parameter defined by (SEN-ID, name, unit)
}
