% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibImpSenData.R
\name{hsGetParID}
\alias{hsGetParID}
\title{hsGetParID}
\usage{
hsGetParID(parName, wqpNames)
}
\arguments{
\item{parName}{parameter name}

\item{wqpNames}{data frame describing (water quality) parameters}
}
\description{
Lookup name of water quality parameter and return its ID or NA if
  the parameter is not yet contained in the data frame \code{wqpNames}
}
