% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibImpSenData.R
\name{hsGetSqls}
\alias{hsGetSqls}
\title{hsGetSqls}
\usage{
hsGetSqls(mdb, tbl, mpID, belowAbove = FALSE, bis2007 = FALSE)
}
\arguments{
\item{mdb}{path to MS Access database}

\item{tbl}{table name}

\item{mpID}{monitoring point ID}

\item{belowAbove}{Default: FALSE}

\item{bis2007}{Default: FALSE}
}
\description{
Returns vector of SQL strings each of which selects the values of one
  water quality parameter (in one column) from table \dQuote{tbl} 
  giving general column names
}
