% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibImpSenData.R
\name{hsGsDataListToMdb}
\alias{hsGsDataListToMdb}
\title{hsGsDataListToMdb}
\usage{
hsGsDataListToMdb(gsDataList, mdb, ...)
}
\arguments{
\item{gsDataList}{List of grab sample data sets}

\item{mdb}{path to MS Access database}

\item{\dots}{arguments passed to hsPutTable}
}
\description{
write table containing all grab sample data in  "all-in-one-table"-format
  to \code{mdb} database
}
