% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibImpSenData.R
\name{hsImpGsData}
\alias{hsImpGsData}
\title{import SENATE's grab sample data from \code{csv}}
\usage{
hsImpGsData(
  csv,
  mdb,
  sep = ",",
  dateFormat = underscoreToPercent("_d/_m/_Y"),
  blockBeginPtrn = "Messstelle",
  tblNamePtrn = "tblSenGrabSmp_@M",
  boolAsk = TRUE,
  dbg = FALSE
)
}
\arguments{
\item{csv}{full path to \code{csv} file}

\item{mdb}{full path to MS Access database (*.mdb)}

\item{sep}{separator in \code{csv} file}

\item{dateFormat}{date format in \code{csv} file (default: \%d/\%m/\%Y)}

\item{blockBeginPtrn}{pattern indicating the begin of a data block in the
file (default: Messstelle)}

\item{tblNamePtrn}{table name pattern. Default: "tblSenGrabSmp_@M" with the
placeholder @M being replaced with the acronym of the monitoring point}

\item{boolAsk}{logical passed to \code{\link{hsWriteBlockToTable}}}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
import SENATE's grab sample data from \code{csv}
}
