% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCalib.R
\name{hsMergeAvailCalibs}
\alias{hsMergeAvailCalibs}
\title{Merge Data in Available Calibrations}
\usage{
hsMergeAvailCalibs(moniPoint, parAcronym, mdbCal, ...)
}
\arguments{
\item{moniPoint}{acronym of monitoring point}

\item{parAcronym}{acronym of parameter}

\item{mdbCal}{path to MS Access database containing calibration data}

\item{\dots}{arguments given to \code{\link{hsPlotEventOverview}}}
}
\value{
data frame with the first two columns representing the timestamp and the
  global calibration, respectively and the following columns representing
  available calibrations, beginning with the "current" calibration that
  is stored in the calibration database under the name 
  <parAcronym>_<moniPoint>, e.g. "AFS_STA"
}
\description{
Merge Data in Available Calibrations
}
