% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDataSource.R
\name{hsMiaDir}
\alias{hsMiaDir}
\title{hsMiaDir}
\usage{
hsMiaDir(dbg = FALSE, ...)
}
\arguments{
\item{dbg}{If \code{TRUE}, debug messages are shown}

\item{\dots}{assignments of the form "name = value"}
}
\description{
Return directory paths containing mdb databases matching the criteria
  defined by the \code{...} parameter list
}
