% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsNextCoefAnaCombi}
\alias{hsNextCoefAnaCombi}
\title{hsNextCoefAnaCombi}
\usage{
hsNextCoefAnaCombi(combi, n)
}
\arguments{
\item{combi}{current combination of which the "successor" combination is to be 
provided}

\item{n}{number of elements in the combination (= length of combination)}
}
\description{
given a current combination the next combination of events as used for 
  coefficient analysis is provided
}
