% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCalib.R
\name{hsPlotCalOverview}
\alias{hsPlotCalOverview}
\title{Plot Results of Calibration Analysis}
\usage{
hsPlotCalOverview(moniPoint, parAcronym, mdbCal, pdf, ...)
}
\arguments{
\item{moniPoint}{acronym of monitoring point}

\item{parAcronym}{acronym of parameter}

\item{mdbCal}{path to MS Access database containing calibration data}

\item{pdf}{path to PDF file to be created}

\item{\dots}{arguments given to \code{\link{hsPlotEventOverview}}}
}
\description{
Plot Results of Calibration Analysis
}
