% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsPlotCoefAnaRes}
\alias{hsPlotCoefAnaRes}
\title{hsPlotCoefAnaRes}
\usage{
hsPlotCoefAnaRes(data, res, recursive = FALSE)
}
\arguments{
\item{data}{\code{data} frame containing columns \emph{tstamp} (time stamp), \emph{pval}
(probe value), \emph{lval} (lab value), \emph{evtID} (event ID)}

\item{res}{result tree as returned by \code{\link{hsCoefAna}}.}

\item{recursive}{Default: FALSE}
}
\description{
Plot function to visualise the regression lines calculated by \code{\link{hsCoefAna}}.
}
