% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCoefAna.R
\name{hsPlotCoefAnaRes2}
\alias{hsPlotCoefAnaRes2}
\title{hsPlotCoefAnaRes2}
\usage{
hsPlotCoefAnaRes2(data, res, olim = NULL, slim = NULL)
}
\arguments{
\item{data}{\code{data} frame containing columns \emph{tstamp} (time stamp), \emph{pval}
(probe value), \emph{lval} (lab value), \emph{evtID} (event ID)}

\item{res}{result tree as returned by \code{\link{hsCoefAna}}.}

\item{olim}{limits of offset values to be used for plotting the offsets}

\item{slim}{limits of slope values to be used for plotting the offsets}
}
\description{
Plot function to visualise the distribution of slopes and offsets of 
  regression lines through possible combinations of events.
}
