% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDataAvailability.R
\name{hsPlotDataAvailability}
\alias{hsPlotDataAvailability}
\title{Plot Data Availability}
\usage{
hsPlotDataAvailability(
  avail,
  colNames = NULL,
  firstIntBeg = NULL,
  lastIntBeg = NULL,
  main = "hsPlotDataAvailability",
  barCols = NULL,
  labelStep = 2,
  firstPlot = TRUE,
  dbg = FALSE,
  ...
)
}
\arguments{
\item{avail}{data frame containing the availibility information as returned by
\code{\link{hsDataAvailability}}}

\item{colNames}{name of column containing availabilities}

\item{firstIntBeg}{timestamp indicating the begin of the first interval to be plotted}

\item{lastIntBeg}{timestamp indicating the begin of the last interval to be plotted}

\item{main}{\code{main} title of barplot}

\item{barCols}{bar colour(s). If \code{avail} is a list of data frames, each data frame is
shown in its own colour as given here in \code{barCols}}

\item{labelStep}{if set to <n>, only every n-th date label will be shown in the plot}

\item{firstPlot}{if TRUE, barplot is replotted, else plot is added to existing plot}

\item{dbg}{If \code{TRUE}, debug messages are shown}

\item{\dots}{further arguments to be passed to R's barplot() function.}
}
\description{
barplot showing data availability in (e.g. daily) time intervals.
}
