% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCalib.R
\name{hsPlotEventOverview}
\alias{hsPlotEventOverview}
\title{Plot Event Overview}
\usage{
hsPlotEventOverview(
  dat,
  evts = NULL,
  evtSepTime = 3600,
  myTitle = "Event Overview",
  plotTypes = rep("l", ncol(dat) - 1),
  dbg = TRUE,
  ...
)
}
\arguments{
\item{dat}{data frame with at least two columns and the timestamps being in the first
column}

\item{evts}{data frame with columns \emph{tBeg} and \emph{tEnd} containing first and
last timestamp, respectively, of the events. If NULL (default) the
events are generated by calling hsEvents (using all timestamps in \code{dat}, 
this is maybe not what we want!!!)}

\item{evtSepTime}{event separation time in seconds, default: 3600 (= 1h)}

\item{myTitle}{plot title}

\item{plotTypes}{vector containing the plot type (cp. type argument of plot function)
for each data column to be plotted}

\item{dbg}{If \code{TRUE}, debug messages are shown}

\item{\dots}{e.g. inset = \code{...}}
}
\description{
Plot Event Overview
}
