% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibGeneral.R
\name{hsPrintToPlot}
\alias{hsPrintToPlot}
\title{hsPrintToPlot}
\usage{
hsPrintToPlot(data, main = "Printed by hsPrintToPlot", addLines = NULL, ...)
}
\arguments{
\item{data}{object to print}

\item{main}{plot title}

\item{addLines}{additional lines}

\item{\dots}{additional arguments passed to legend, e.g. cex}
}
\description{
prints content of an object to a plot
}
