% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSelectTimeseriesInterval.R
\name{hsSel}
\alias{hsSel}
\title{Filter Timeseries Interactively}
\usage{
hsSel(tseries, interval = FALSE, tsfield = names(tseries)[1], dbg = FALSE)
}
\arguments{
\item{tseries}{data frame containing time series data}

\item{interval}{logical. Default: FALSE}

\item{tsfield}{Name of column containing the time stamps. Default: name of 
first column}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
Filter timeseries interactively for \code{interval} or parallel intervals
}
