% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSelectTimeseriesInterval.R
\name{hsSelectParallelTimeIntervals}
\alias{hsSelectParallelTimeIntervals}
\title{Interactively Select Parallel Time Intervals}
\usage{
hsSelectParallelTimeIntervals(tseries, tsfield = names(tseries)[1])
}
\arguments{
\item{tseries}{data.frame containing timeseries (at least one timestamp column and one
additional numeric column)}

\item{tsfield}{name of timestamp column; default: name of first column in \code{tseries}}
}
\description{
Interactively select parallel time intervals from data.frame containing timeseries
}
