% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSelectTimeseriesInterval.R
\name{hsSelectTimeInterval}
\alias{hsSelectTimeInterval}
\title{Interactively Select Time Interval}
\usage{
hsSelectTimeInterval(tseries, tsfield = names(tseries)[1])
}
\arguments{
\item{tseries}{data.frame containing timeseries (at least one timestamp column and one
additional numeric column)}

\item{tsfield}{name of timestamp column; default: name of first column in \code{tseries}}
}
\description{
Interactively select time interval from data.frame containing timeseries
}
