% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibDataSource.R
\name{hsSourceList}
\alias{hsSourceList}
\title{hsSourceList}
\usage{
hsSourceList(keyptrn = "^DB_", dbg = FALSE, ...)
}
\arguments{
\item{keyptrn}{Pattern matching the keys representing databases. 
Default: "^DB_"}

\item{dbg}{If \code{TRUE}, debug messages are shown}

\item{\dots}{definition of filter criteria given as "key = value" pairs}
}
\description{
Return data frame containing ids, properties and paths of mdb databases
  matching criteria given in \code{...} argument list
}
