% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCalib.R
\name{hsSpecCal}
\alias{hsSpecCal}
\title{Special calibration}
\usage{
hsSpecCal(moniPoint, parAcronym, calName = NULL, mdbCal = NULL)
}
\arguments{
\item{moniPoint}{acronym of monitoring point}

\item{parAcronym}{acronym of parameter}

\item{calName}{name of calibration}

\item{mdbCal}{Path to database containing queries that getting calibrated
data according to the currently active calibration setting}
}
\description{
Get calibrated data according to a special calibration given by its name.
  Instead of running the prepared query in \dQuote{KWB_CAL.mdb} the specific
  SQL query respecting the given calibratio name is built and run here.
}
