% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCalib.R
\name{hsSqlExCal}
\alias{hsSqlExCal}
\title{SQL to Get Calibrated Data}
\usage{
hsSqlExCal(moniPoint, parAcronym, calName = NULL)
}
\arguments{
\item{moniPoint}{acronym of monitoring point}

\item{parAcronym}{acronym of parameter}

\item{calName}{name of calibration}
}
\description{
Generate SQL expression needed to get calibrated data
}
