% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibCalib.R
\name{hsTestCombi}
\alias{hsTestCombi}
\title{Test Combination of Events}
\usage{
hsTestCombi(
  data,
  combi,
  plot = TRUE,
  plot.main = "Lab values vs probe values",
  COLS = rainbow(length(unique(data$evtid)))
)
}
\arguments{
\item{data}{data frame}

\item{combi}{combination of events used for calibration}

\item{plot}{logical indicating whether to plot}

\item{plot.main}{plot title}

\item{COLS}{vector of colours}
}
\description{
Evaluate a specific combination of events used for calibration
}
