% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibTimeshift.R
\name{hsTimeshift}
\alias{hsTimeshift}
\title{Timeshift}
\usage{
hsTimeshift(
  hq,
  threshold,
  upstream = TRUE,
  tsField = names(hq)[1],
  valField = names(hq)[2],
  quaFields = names(hq)[-c(1, 2)],
  maxTDiff = 3600,
  valFactor = 1,
  dbg = FALSE
)
}
\arguments{
\item{hq}{hydraulic and water quality data in one data frame}

\item{threshold}{Threshold that shall be reached/exceeded by the sum of
successive values in column \emph{valField} of which the maximum time
difference is below or equal \emph{maxTDiff}.}

\item{upstream}{if TRUE, the algorithm \dQuote{looks} \code{upstream}, else
downstream}

\item{tsField}{name of timestamp field in \emph{hq}}

\item{valField}{name of column in \emph{hq} containing the values to be
summed up until the \code{threshold} is reached}

\item{quaFields}{vector containing column names of water quality parameters,
e.g. c("AFS", "CSB", "CSBf")}

\item{maxTDiff}{Maximum allowed time difference in seconds between two
related timestamps.}

\item{valFactor}{factor to be applied to column \emph{valField} before
calculating value sums.}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
\code{upstream} or downstream "timeshift" of water quality data given
time-series of hydraulic and water quality data in one data frame
}
