% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibTimeshift.R
\name{hsTimeshiftPlot}
\alias{hsTimeshiftPlot}
\title{Timeshift Plot}
\usage{
hsTimeshiftPlot(
  frmOrig,
  frmUs,
  frmDs,
  evt,
  fieldNames,
  i,
  evtDurFract,
  type1,
  boolOverview = FALSE,
  fieldPrefix = ""
)
}
\arguments{
\item{frmOrig}{data frame with original measurements}

\item{frmUs}{data frame with measurements time-shifted upstream}

\item{frmDs}{data frame with measurements time-shifted downstream}

\item{evt}{event list}

\item{fieldNames}{field names}

\item{i}{index}

\item{evtDurFract}{event duration fraction}

\item{type1}{plot type, e.g. "p" for "points", passed to \code{\link{plot}}}

\item{boolOverview}{should an overview be plotted?}

\item{fieldPrefix}{field prefix}
}
\description{
A plot is generated, containing flow measurements as well as measurements
(original, upstream/downstream time-shifted) of one water quality parameter
and for one overflow events \code{i} contained in the event list "evt".
}
