% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibGeneral.R
\name{hsUnionSqls}
\alias{hsUnionSqls}
\title{hsUnionSqls}
\usage{
hsUnionSqls(sqls, maxUnions)
}
\arguments{
\item{sqls}{vector of SQL statements}

\item{maxUnions}{number of maximum UNIONs allowed}
}
\value{
vector of UNION queries
}
\description{
Creates UNION-queries of given SQL queries, respecting the maximum number
  of subqueries to be used in one and the same UNION query.
}
