% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibImpSenData.R
\name{hsWriteBlockToTable}
\alias{hsWriteBlockToTable}
\title{write data block to database table}
\usage{
hsWriteBlockToTable(
  channel,
  block,
  blockName,
  tblNamePtrn = "tblSenGrabSmp_@M",
  boolAsk,
  dbg = FALSE
)
}
\arguments{
\item{channel}{database connection}

\item{block}{block}

\item{blockName}{block name}

\item{tblNamePtrn}{name of table to be written in database. @M is replaced with the name 
of the monitoring point}

\item{boolAsk}{logical. Ask before tropping existing tables?}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
write data \code{block} to database table
}
