% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{na_checksum}
\alias{na_checksum}
\title{number representing combination of gauges with NA-failure}
\usage{
na_checksum(x)
}
\arguments{
\item{x}{data frame or matrix}
}
\description{
calculate checksum for each row of a data frame or matrix \emph{x}. 
  Each combination of NA-occurrence in the row gets a different checksum.
  Therefore, each column of \emph{x} is represented as a power of two
  and for the columns in which the row contains NA the corresponding powers
  of two are added.
}
