% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibGeneral.R
\name{pageAndPlot}
\alias{pageAndPlot}
\title{output object to plots of same row number}
\usage{
pageAndPlot(data, rpp = 60, to.pdf = TRUE, ...)
}
\arguments{
\item{data}{\code{data} frame to plot to pdf}

\item{rpp}{rows per page}

\item{to.pdf}{if TRUE (default) the output goes into a temporary PDF file, otherwise
to the standard plot device}

\item{\dots}{arguments to be passed to \code{\link{hsPrintToPlot}}, e.g. main, cex}
}
\description{
capture the output of printing an object, split this output into blocks of 
  equal size (row per page) and print these blocks as plots using
  \code{\link{hsPrintToPlot}}
}
