% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_given_rain_events}
\alias{plot_given_rain_events}
\title{plot rain events from rain data and event info}
\usage{
plot_given_rain_events(
  rain,
  evt,
  imin = 1,
  imax = -1,
  sigWidth_s,
  myCex = 0.7,
  myMinN
)
}
\arguments{
\item{rain}{data frame containing \code{rain} data}

\item{evt}{event information as returned by \code{kwb.event::hsEvents}}

\item{imin}{row index in \emph{evt} of first event to be considered. Default: 1}

\item{imax}{row index in \emph{evt} of last event to be considered. Set to -1 
(default) to consider all events.}

\item{sigWidth_s}{\code{rain} signal width in seconds}

\item{myCex}{default: 0.7}

\item{myMinN}{minimum total precipitation in mm. Rain events with less precipitation are
not plotted}
}
\description{
plot all \code{rain} events from given \code{rain} data and event information. 
  This function calls \code{\link{plot_one_rain_event_2}} in a loop over all events
  with indices between \emph{imin} and \emph{imax}.
}
