% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_one_rain_event}
\alias{plot_one_rain_event}
\title{Plot One Rain Event}
\usage{
plot_one_rain_event(frmR, evt, i)
}
\arguments{
\item{frmR}{data frame containing rain data}

\item{evt}{data frame containing event data}

\item{i}{index into event data}
}
\description{
Plot One Rain Event
}
