% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_one_rain_event_2}
\alias{plot_one_rain_event_2}
\title{Plot One Rain Event}
\usage{
plot_one_rain_event_2(
  rain,
  rainEventIndex = 1,
  sigWidth_s = NULL,
  boolBarplot = TRUE,
  myCex = 1,
  myMinN = 0
)
}
\arguments{
\item{rain}{data frame containing rain data}

\item{rainEventIndex}{rain event index}

\item{sigWidth_s}{\code{rain} signal width in seconds}

\item{boolBarplot}{logical. Default: \code{TRUE}}

\item{myCex}{character expansion factor. Default: 1}

\item{myMinN}{minimum total precipitation. Default: 0}
}
\description{
plot one \code{rain} event by calling \code{\link{plot_one_rain_gauge_event}}
}
