% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_one_rain_event_3}
\alias{plot_one_rain_event_3}
\title{Plot One Rain Event 3}
\usage{
plot_one_rain_event_3(frmR, evt, i, myCols)
}
\arguments{
\item{frmR}{data frame containing rain data}

\item{evt}{data frame containing rain event data}

\item{i}{index}

\item{myCols}{vector of colours}
}
\description{
Plot One Rain Event 3
}
