% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_one_rain_gauge_event}
\alias{plot_one_rain_gauge_event}
\title{Plot Rain Event}
\usage{
plot_one_rain_gauge_event(
  rain,
  sigWidth_s = NULL,
  i = 2,
  Nmax = 1,
  boolLabels = FALSE,
  boolDebug = FALSE,
  myBlue = rainbow(8)[6],
  nMaxLabels = 60,
  dbg = FALSE
)
}
\arguments{
\item{rain}{data frame containing rain data}

\item{sigWidth_s}{signal width in seconds}

\item{i}{Default: 2}

\item{Nmax}{Default: 1}

\item{boolLabels}{plot labels?}

\item{boolDebug}{logical}

\item{myBlue}{nice blue colour}

\item{nMaxLabels}{# max number of labels per page}

\item{dbg}{If \code{TRUE}, debug messages are shown}
}
\description{
Plot Rain Event
}
