% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_rain_events}
\alias{plot_rain_events}
\title{plot rain events from given rain data}
\usage{
plot_rain_events(
  rain,
  strTimestamp = "Zeitstempel",
  strPdf = NULL,
  irng = c(1, -1),
  myCex = 0.7,
  myMinN,
  evtSepTime_s = 60 * 60 * 6,
  sigWidth_s = 300,
  events = TRUE
)
}
\arguments{
\item{rain}{data frame containing \code{rain} data}

\item{strTimestamp}{name of timestamp field, default: "Zeitstempel"}

\item{strPdf}{optional. path to output pdf file}

\item{irng}{vector of two integer values determining the first and last index,
respectively, in the data frame of \code{events} (created within this function) 
to be plotted. Second element can be -1 to indicate the index of the
last event. Default: c(1, -1)}

\item{myCex}{default: 0.7}

\item{myMinN}{minimum precipitation in mm. Events with a total depth less than this 
value are not plotted}

\item{evtSepTime_s}{event separation time in seconds. Default: 6*3600 = 6 hours}

\item{sigWidth_s}{\code{rain} signal width in seconds. Default: 300 = 5 minutes}

\item{events}{one plot per event (\code{events} = TRUE) or one plot per day (\code{events} = FALSE)?
default: TRUE}
}
\description{
plot \code{rain} \code{events} from given \code{rain} data. Either one plot per day or one plot
  per \code{rain} event (created within this function). This function calls 
  \code{\link{plot_given_rain_events}}
}
