% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_rain_events_to_pdf}
\alias{plot_rain_events_to_pdf}
\title{Plot Rain Events to PDF}
\usage{
plot_rain_events_to_pdf(frmR, evt, strPdf, myCols, myVersion = 1)
}
\arguments{
\item{frmR}{data frame containing rain data}

\item{evt}{data frame containing rain event data}

\item{strPdf}{path to PDF file}

\item{myCols}{vector of colours}

\item{myVersion}{Default: 1}
}
\description{
Plot Rain Events to PDF
}
