% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_rain_subevents_by_checksum}
\alias{plot_rain_subevents_by_checksum}
\title{Plot Rain Subevents by Checksum}
\usage{
plot_rain_subevents_by_checksum(rain, evt, n = nrow(evt))
}
\arguments{
\item{rain}{data frame containing rain data}

\item{evt}{data frame containin rain event data}

\item{n}{Default: number of rows in \code{evt}}
}
\description{
Plot Rain Subevents by Checksum
}
