% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plot_rain_with_subevent_lines}
\alias{plot_rain_with_subevent_lines}
\title{Plot Rain with Subevent Lines}
\usage{
plot_rain_with_subevent_lines(rain, subevt, rainEventIndex)
}
\arguments{
\item{rain}{data frame containing rain data}

\item{subevt}{subevent data}

\item{rainEventIndex}{rain event index}
}
\description{
Plot Rain with Subevent Lines
}
