% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plotDataAvailability}
\alias{plotDataAvailability}
\title{Plot Data Availability}
\usage{
plotDataAvailability(
  rainStat,
  n = nrow(rainStat),
  main = "Data Availability",
  cex = 0.9,
  barColours = c("lightgreen", "indianred")
)
}
\arguments{
\item{rainStat}{rain statistics with one row per rain event}

\item{n}{number of rain event}

\item{main}{plot title}

\item{cex}{character expansion factor. Default: 0.9}

\item{barColours}{vector of bar colours}
}
\description{
Plot Data Availability
}
