% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{plotTotalPrecipitation}
\alias{plotTotalPrecipitation}
\title{Plot Total Precipitation}
\usage{
plotTotalPrecipitation(
  rainStat,
  n = nrow(rainStat),
  main = "Total Precipitation",
  Nmax_mm = NA,
  cex = 0.9
)
}
\arguments{
\item{rainStat}{data frame with column \emph{N_mm} and row names indicating
rain gauge names}

\item{n}{number of rain series}

\item{main}{plot title}

\item{Nmax_mm}{maximum total precipitation in mm}

\item{cex}{character expansion factor. Default: 0.9}
}
\description{
Plot Total Precipitation
}
