% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{prepare_rain_data}
\alias{prepare_rain_data}
\title{Prepare Rain Data}
\usage{
prepare_rain_data(
  rain,
  myVersion,
  event_sep_time = 60 * 60 * 6,
  signal_width = 300
)
}
\arguments{
\item{rain}{rain data}

\item{myVersion}{1 or 2, depending on required output}

\item{event_sep_time}{event separation time in seconds}

\item{signal_width}{signal width, i.e. time period that a rain data point
represents, in seconds}
}
\description{
Prepare Rain Data
}
