% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsLibRainDist.R
\name{read_rain_from_mdb}
\alias{read_rain_from_mdb}
\title{Read Rain Data from Database}
\usage{
read_rain_from_mdb(myVersion, strDb)
}
\arguments{
\item{myVersion}{1 (to read from tbl_Regen_alleEZG_05min) or 2 (to read from
tbl_Regen_alleEZG_05min)}

\item{strDb}{path to MS Access database}
}
\description{
Read Rain Data from Database
}
